Update README.md
PRAGMATICA : Journal of Linguistics and Literature is the Journal who published research articles and of theoretical articles in linguistics science. The scope of this journal includes linguistics micro, linguistics macro, and literature. The scope of linguistics micro restricted to the fields of syntax, morphology, phonology, typology and semantics, whereas the scope of linguistics macro restricted to sociolinguistics, antropolinguistics, language-teaching, eco-linguistic and translation, for the scope of literature restricted to linguistics aspects contained in English literature. The journal is published twice a year every May and November published by CV Tirta Pustaka Press.